Replace Vite to Create-React-App
