+++ src/components/NavBar.js
@@ -1,12 +1,12 @@
-import { useState } from "react";
-import { ImSearch } from "react-icons/im";
-import { useDispatch } from "react-redux";
-import styles from "../CSS/NavBar.module.css";
-import { setSearchKey } from "../redux/Coins/SearchSlice";
+import { useState } from 'react';
+import { ImSearch } from 'react-icons/im';
+import { useDispatch } from 'react-redux';
+import styles from '../CSS/NavBar.module.css';
+import { setSearchKey } from '../redux/Coins/SearchSlice';
 
 const Navbar = () => {
   const dispatch = useDispatch();
-  const [input, setinput] = useState("");
+  const [input, setinput] = useState('');
 
   const onchange = (event) => {
     setinput(event.target.value);
@@ -16,17 +16,12 @@ const Navbar = () => {
   const onsubmit = (event) => {
     dispatch(setSearchKey(input));
     event.preventDefault();
-    console.log(input);
-    setinput("");
-  };
-
-  const onlogo = () => {
-   location.reload();
+    setinput('');
   };
 
   return (
     <nav className={styles.nav}>
-      <h2 className={styles.h2} onClick={onlogo}>
+      <h2 className={styles.h2}>
         CryptoPlace
       </h2>
       <form className={styles.form} onSubmit={onsubmit}>
@@ -35,7 +30,7 @@ const Navbar = () => {
           placeholder="   search"
           value={input}
           onChange={onchange}
-        ></input>
+        />
         <button type="submit" className={styles.btn}>
           <ImSearch className={styles.search} />
         </button>
