+++ src/redux/Coins/CoinSlice.js
@@ -1,19 +1,17 @@
-import { createSlice } from "@reduxjs/toolkit";
-import { createAsyncThunk } from "@reduxjs/toolkit";
-import axios from "axios";
+import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 
-const url =
-  "https://api.coingecko.com/api/v3/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=60&page=1&sparkline=false&price_change_percentage=1h%2C24h%2C7d&locale=en";
-const FetchData = createAsyncThunk("Fetch Data", async () => {
-  try{
-     const response = await axios.get(url);
-  const data = response.data;
-  // console.log(data);
-  return data
+import axios from 'axios';
+
+const url = 'https://api.coingecko.com/api/v3/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=60&page=1&sparkline=false&price_change_percentage=1h%2C24h%2C7d&locale=en';
+const FetchData = createAsyncThunk('Fetch Data', async () => {
+  try {
+    const response = await axios.get(url);
+    const { data } = response;
+    // console.log(data);
+    return data;
+  } catch (error) {
+    throw new Error('Something went wrong! :', error);
   }
- catch(error){
-  throw new Error('Something went wrong! :', error);
- }
 });
 
 FetchData();
@@ -48,5 +46,5 @@ const CoinSlice = createSlice({
   },
 });
 
-export default CoinSlice.reducer ;
-export {FetchData};
+export default CoinSlice.reducer;
+export { FetchData };
